WIP Moving swift file checks to rpc_support role
This probably isn't the right place for these tasks to live, but
for right now I don't know of a better place.

These changes create a file in a swift container to be accessed by
monitoring checks. A successful return of the file means swift is
completely functional from the proxy through the ring to the backing
storage device. This gives a more correct status of swift's
availability. Any result other than 200 is considered a failure.

Add tasks to create Keystone user
Add the user to the swiftoperators role
Add tasks to upload an index file into a check-specific container
Set permission on the container to be world readable
Set the index file to be the container default index
Fetch the container token
Fact set a URL to be checked to access the index file