Catch up commit
Commit the latest work.
This is the result of a lot of experimenting and testing. The simplest
way to get authenticated and call the appropriate commands seems to be
to source an openrc file right now. But I feel like there must be a
better way. I plan to survey the group at a later day to potentially
change this pattern.

I've added a secret for the keystone password for the file check user,
added all the necessary tasks to get swift configured properly, and the
variables required to make the play basically function. I still need to
get a decent way to push or create an index file onto a remote host,
then upload that file.